requirements.txt
PyYAML==5.3.1
supervisely==6.1.*
htmllistparse==0.5.2